refactor(preferences): move preference route into its own controller/view
