+++ mix.exs
@@ -39,6 +39,7 @@ defmodule EpochtalkServer.MixProject do
       {:ecto_sql, "~> 3.6"},
       {:ex_doc, "~> 0.28.5"},
       {:gen_smtp, "~> 1.2"},
+      {:guardian, "~> 2.2"},
       {:guardian_phoenix, "~> 2.0"},
       {:guardian_db, "~> 2.1"},
       {:guardian_redis, "~> 0.1"},
