+++ test/epochtalk_server_web/views/preference_view_test.exs
@@ -1,0 +1,7 @@
+defmodule EpochtalkServerWeb.PreferenceViewTest do
+  use EpochtalkServerWeb.ConnCase, async: true
+  alias EpochtalkServerWeb.PreferenceView
+
+  # Specify that we want to use doctests:
+  doctest PreferenceView
+end
